// src/app/Models/GetFlowViewAttendUnusualDataClass.ts
// Generated by https://quicktype.io

export interface GetFlowViewAttendUnusualDataClass {
    EmpID:             string;
    EmpName:           string;
    DeptName:          string;
    Date:              string;
    EliminateLate:     boolean;
    EliminateEarly:    boolean;
    EliminateOnBefore: boolean;
    EliminateOffAfter: boolean;
    EliminateAbsent:   boolean;
    RoteNameC:         string;
    ErrorState:        string;
    Key:               string;
    ProcessFlowID:     number;
    FormName:          string;
    AppEmpID:          string;
    AppEmpName:        string;
    AppDeptName:       string;
    ManageEmpName:     string;
    RealManageEmpName: string;
    State:             string;
    Handle:            boolean;
    Take:              boolean;
    TransSign:         boolean;
}